fix: changes to uint

+++ models/taxonomies/category.go
@@ -10,24 +10,24 @@ type Category struct {
 
 type SubCategory struct {
 	gorm.Model
-	Id         int64    `json:"id"`
+	Id         uint     `json:"id"`
 	Name       string   `json:"name"`
-	CategoryId int64    `json:"category_id"`
+	CategoryId uint     `json:"category_id"`
 	Category   Category `json:"category"`
 }
 
 type Segment struct {
 	gorm.Model
-	Id            int64       `json:"id"`
+	Id            uint        `json:"id"`
 	Name          string      `json:"name"`
-	SubCategoryId int64       `json:"sub_category_id"`
+	SubCategoryId uint        `json:"sub_category_id"`
 	SubCategory   SubCategory `json:"sub_category"`
 }
 
 type SubSegment struct {
 	gorm.Model
-	Id        int64   `json:"id"`
+	Id        uint    `json:"id"`
 	Name      string  `json:"name"`
-	SegmentId int64   `json:"segment_id"`
+	SegmentId uint    `json:"segment_id"`
 	Segment   Segment `json:"segment"`
 }
